+++ vlasiator.cpp
@@ -36,7 +36,7 @@
 
 #include <fsgrid.hpp>
 
-#include "vlasovmover.h"
+#include "vlasovsolver/vlasovmover.h"
 #include "vlasovsolver/vec.h"
 #include "definitions.h"
 #include "mpiconversion.h"
